requirements.txt
pandas
nltk
streamlit
feedparser
python-dotenv
google-api-python-client
praw
